rework metrics
