+++ src/plugins/history/Metrics/MetricsFull.vue
@@ -1,7 +1,7 @@
 <script lang="ts">
 import get from 'lodash/get';
 import parseDuration from 'parse-duration';
-import { Component } from 'vue-property-decorator';
+import { Component, Prop } from 'vue-property-decorator';
 
 import { defaultLabel, targetBuilder, targetSplitter } from '@/components/Graph/functional';
 import CrudComponent from '@/components/Widget/CrudComponent';
@@ -12,10 +12,13 @@ import { DEFAULT_DECIMALS, DEFAULT_FRESH_DURATION } from './getters';
 import { MetricsConfig } from './types';
 
 @Component
-export default class MetricsForm extends CrudComponent {
+export default class MetricsFull extends CrudComponent {
   DEFAULT_FRESH_DURATION = DEFAULT_FRESH_DURATION;
   parseDuration = parseDuration;
   durationString = durationString;
+
+  @Prop({ type: Boolean, default: false })
+  public readonly inDialog!: boolean;
 
   get config(): MetricsConfig {
     return {
@@ -26,6 +29,14 @@ export default class MetricsForm extends CrudComponent {
       decimals: {},
       ...this.widget.config,
     };
+  }
+
+  get scrollParentClass(): Mapped<boolean> {
+    return { 'col-grow': true, 'scroll-parent': this.inDialog };
+  }
+
+  get selectorParent(): string {
+    return this.inDialog ? 'q-scroll-area' : 'div';
   }
 
   get selected(): string[] {
@@ -81,29 +92,30 @@ export default class MetricsForm extends CrudComponent {
 </script>
 
 <template>
-  <q-card dark class="widget-modal">
-    <WidgetDialogToolbar :crud="crud" />
+  <q-card dark v-bind="$attrs">
+    <slot name="toolbar" />
+    <slot name="warnings" />
 
     <q-card-section>
       <q-expansion-item default-opened group="modal" icon="mdi-file-tree" label="Metrics">
-        <div class="scroll-parent">
-          <q-scroll-area>
+        <div :class="scrollParentClass">
+          <component :is="selectorParent">
             <MetricSelector :selected.sync="selected" />
-          </q-scroll-area>
+          </component>
         </div>
       </q-expansion-item>
 
       <q-expansion-item group="modal" icon="mdi-tag-multiple" label="Labels">
-        <div class="scroll-parent">
-          <q-scroll-area>
+        <div :class="scrollParentClass">
+          <component :is="selectorParent">
             <LabelSelector :selected="selected" :renames.sync="renames" />
-          </q-scroll-area>
+          </component>
         </div>
       </q-expansion-item>
 
       <q-expansion-item group="modal" icon="warning" label="Old data warnings">
-        <div class="scroll-parent">
-          <q-scroll-area>
+        <div :class="scrollParentClass">
+          <component :is="selectorParent">
             <q-item dark>
               <q-item-section>Metric</q-item-section>
               <q-item-section>Warn when older than</q-item-section>
@@ -123,13 +135,13 @@ export default class MetricsForm extends CrudComponent {
                 <q-btn icon="restore" flat @click="resetFreshDuration(field)" />
               </q-item-section>
             </q-item>
-          </q-scroll-area>
+          </component>
         </div>
       </q-expansion-item>
 
       <q-expansion-item group="modal" icon="mdi-numeric" label="Rounding">
-        <div class="scroll-parent">
-          <q-scroll-area>
+        <div :class="scrollParentClass">
+          <component :is="selectorParent">
             <q-item dark>
               <q-item-section>Metric</q-item-section>
               <q-item-section>Number of decimals</q-item-section>
@@ -152,7 +164,7 @@ export default class MetricsForm extends CrudComponent {
                 <q-btn icon="mdi-undo-variant" flat @click="resetDecimals(field)" />
               </q-item-section>
             </q-item>
-          </q-scroll-area>
+          </component>
         </div>
       </q-expansion-item>
     </q-card-section>
